Add Contact page. Rename components filenames
